ci: Build  missing bullseye debian-wx32.

+++ ci/circleci-build-debian-docker.sh
@@ -45,6 +45,10 @@ else
     apt-get install -y cmake
 fi
 
+if [ -n "@BUILD_WX32@" ]; then
+  exit 1
+fi
+
 
 cd /ci-source
 rm -rf build-debian; mkdir build-debian; cd build-debian
@@ -55,6 +59,7 @@ chown --reference=.. .
 EOF
 
 sed -i "s/@TARGET_TUPLE@/$TARGET_TUPLE/" $ci_source/build.sh
+sed -i "s/@BUILD_WX32@/$BUILD_WX32/" $ci_source/build.sh
 
 
 # Run script in docker image
